requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-chroma>=0.2.0
chromadb>=0.6.0
python-multipart>=0.0.20
